genai/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.3.1
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pluggy==1.5.0
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.5
ruff==0.11.8
setuptools==80.4.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.5
weaviate-client==4.14.1
websockets==15.0.1
